pyproject.toml
[project]
name = "vllm-workload-load-test"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "matplotlib>=3.7",
    "numpy>=1.24",
    "pandas>=2.0",
    "pydantic>=2.0",
    "pyyaml>=6.0",
    "seaborn>=0.12",
    "vllm>=0.11.0",
]

[dependency-groups]
dev = [
    "ruff>=0.14.5",
]
